[MultiverseAction] Made remove_object_from_simulator return a boolean value indicating whether the simulator removed the object or not and if true the world would remove the object from its list of objects else it will leave it be.
