+++ src/pycram/worlds/multiverse_datastructures/dataclasses.py
@@ -4,6 +4,18 @@ from typing_extensions import List
 
 
 @dataclass
+class MultiverseMetaData:
+    """Meta data for the Multiverse Client, the simulation_name should be non-empty and unique for each simulation"""
+    world_name: str = "world"
+    simulation_name: str = "cram"
+    length_unit: str = "m"
+    angle_unit: str = "rad"
+    mass_unit: str = "kg"
+    time_unit: str = "s"
+    handedness: str = "rhs"
+
+
+@dataclass
 class RayResult:
     """
     A dataclass to store the ray result. The ray result contains the body name that the ray intersects with and the
